+++ src/main/java/com/joorani/myshop/entity/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.joorani.myshop.product.domain.entity;
+package com.joorani.myshop.entity;
 
 import lombok.Getter;
 
@@ -7,7 +7,7 @@ public enum ProductStatus {
     ON_SALE("판매중"),
     OUT_OF_STOCK("품절");
 
-    private String desc;
+    private final String desc;
 
     ProductStatus(String desc) {
         this.desc = desc;
