--- src/main/java/com/joorani/myshop/member/domain/entity/Role.java
@@ -1,5 +1,0 @@
-package com.joorani.myshop.member.domain.entity;
-
-public enum Role {
-    SELLER, BUYER, ADMIN
-}
