+++ src/main/java/com/joorani/myshop/repository/StoreRepository.java
@@ -1,6 +1,6 @@
-package com.joorani.myshop.member.domain.repository;
+package com.joorani.myshop.repository;
 
-import com.joorani.myshop.member.domain.entity.Store;
+import com.joorani.myshop.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
